Move LeasableMachine and MachineSet under internal/machine/
